+++ cmd/products/products_apply.go
@@ -1,17 +1,22 @@
 package products
 
 import (
+	"errors"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"os"
+	"path/filepath"
 
-	"gopkg.in/yaml.v3"
+	"gopkg.in/yaml.v2"
 
 	"github.com/ecommerce-builder/ecom-cli-tool/configmgr"
 	"github.com/ecommerce-builder/ecom-cli-tool/eclient"
 	"github.com/spf13/cobra"
 )
+
+type productContainer struct {
+	Product eclient.ProductApply `yaml:"product"`
+}
 
 // NewCmdProductsApply returns new initialized instance of the apply sub command
 func NewCmdProductsApply() *cobra.Command {
@@ -22,7 +27,7 @@ func NewCmdProductsApply() *cobra.Command {
 	}
 
 	var cmd = &cobra.Command{
-		Use:   "apply <product.yaml>",
+		Use:   "apply <product.yaml>|<dir>",
 		Short: "Create or update an exising product",
 		Args:  cobra.ExactArgs(1),
 		Run: func(cmd *cobra.Command, args []string) {
@@ -32,49 +37,80 @@ func NewCmdProductsApply() *cobra.Command {
 				log.Fatal(err)
 			}
 
-			data, err := ioutil.ReadFile(args[0])
+			isDir, err := isDirectory(args[0])
+			if err != nil {
+				fmt.Fprintf(os.Stderr, "%v\n", err)
+				os.Exit(1)
+			}
+			if !isDir {
+				if err := applyProduct(client, args[0]); err != nil {
+					if err == errMissingEAN {
+						fmt.Fprintf(os.Stderr, "Skipping %s as EAN is missing\n", args[0])
+						os.Exit(1)
+					}
+					fmt.Fprintf(os.Stderr, "%v\n", err)
+					os.Exit(1)
+				}
+				os.Exit(0)
+			}
+
+			matches, err := filepath.Glob(args[0] + "/*.yaml")
 			if err != nil {
 				fmt.Fprintf(os.Stderr, "%v\n", err)
 				os.Exit(1)
 			}
 
-			var p struct {
-				Product eclient.Product `yaml:"product"`
-			}
-			err = yaml.Unmarshal([]byte(data), &p)
-			if err != nil {
-				log.Fatalf("error: %v", err)
-			}
-			exists, err := client.ProductExists(p.Product.SKU)
-			if err != nil {
-				log.Fatalf("error: %+v", err)
-			}
-			pc := eclient.ProductCreate{
-				SKU:  p.Product.SKU,
-				EAN:  p.Product.EAN,
-				URL:  p.Product.URL,
-				Name: p.Product.Name,
-				Data: p.Product.Data,
-			}
-			if !exists {
-				_, err = client.CreateProduct(&pc)
-				if err != nil {
-					log.Fatalf("create product failed sku=%q: %+v", p.Product.SKU, err)
+			for _, file := range matches {
+				if err := applyProduct(client, file); err != nil {
+					if err == errMissingEAN {
+						fmt.Fprintf(os.Stderr, "Skipping %s as EAN is missing\n", file)
+						continue
+					}
+					fmt.Fprintf(os.Stderr, "%v\n", err)
+					os.Exit(1)
 				}
-				os.Exit(0)
-			}
-			pu := eclient.ProductUpdate{
-				EAN:  p.Product.EAN,
-				URL:  p.Product.URL,
-				Name: p.Product.Name,
-				Data: p.Product.Data,
-			}
-			_, err = client.UpdateProduct(p.Product.SKU, &pu)
-			if err != nil {
-				log.Fatalf("update product failed sku=%q: %+v", p.Product.SKU, err)
 			}
 			os.Exit(0)
 		},
 	}
 	return cmd
+}
+
+var errMissingEAN = errors.New("missing EAN")
+
+func applyProduct(ec *eclient.EcomClient, filen string) error {
+	file, err := os.Open(filen)
+	if err != nil {
+		return err
+	}
+	defer file.Close()
+	p := productContainer{}
+	dec := yaml.NewDecoder(file)
+	if err := dec.Decode(&p); err != nil {
+		return err
+	}
+	//if p.Product.EAN == "" {
+	//	return errMissingEAN
+	//}
+	pu := eclient.ProductApply{
+		EAN:     p.Product.EAN,
+		Path:    p.Product.Path,
+		Name:    p.Product.Name,
+		Images:  p.Product.Images,
+		Pricing: p.Product.Pricing,
+		Content: p.Product.Content,
+	}
+	_, err = ec.ReplaceProduct(p.Product.SKU, &pu)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+func isDirectory(path string) (bool, error) {
+	fileInfo, err := os.Stat(path)
+	if err != nil {
+		return false, err
+	}
+	return fileInfo.IsDir(), err
 }